feat: login scaffold
